Fix whole shares before fractional share debacle.
Found a bug in whole share priority before fractional.
After aggregation of shares for buy or sell we want to trim fractionals
if there is a whole share and place the whole share as part of one workflow.
This logic had a flaw of resulting in ordering with 0 shares
shares_to_buy % 1 > 0 and then order int(shares_to_buy).

Adusted integration tests accordingly

Started adding docstring for integration tests.