Current channel structure in utils doesn't suite different bots. 1 price channel, multiple bots (#17)
* Current channel structure in utils doesn't suite different bots.
To simplify the channel streams for each bot, we've separated it by
ticker and we have 1 registration channel for the broker to subscribe
to once a bot comes up.

* Alpaca broker goes redissing. LoL. (#18)

Subscribe to Registration channel. Use the messages to stop streamers.
Subscribe to websocket for existing and new ticker.

Right now we receive order update and price update per ticker.
One know problem is order updates for unregisterd tickers will go to
redis as well. Only producable from frontend alpaca. Not a big deal